requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.11,<2.0
Django>=1.4.2,<1.8
oslo.config>=1.11.0  # Apache-2.0
python-keystoneclient>=1.3.0
six>=1.9.0
# Python Imaging Library - PIL is almost dead.
# Pillow is a PIL upgraded edition.
Pillow
